Everything seems to be working. I need to modify the grade_df dataframe to add some more students and grades to more fully demonstrate the formatting functions.
I need to add an option to use existing formatting features that do not conflict with new ones. That is, if there is a border in the existing cell and the new format does not specify a border, the existing border would remain.This is not trivial, because it is a cell by cell formatting command.